local: toolchain_util: Set COREOS_SDK_VERSION to COREOS_VERSION
This is a work-around.

FIXME: To build packages with the jenkins built sdk tools the
jenkins built sdk packages need to be put into the chroot.  To
do that a --getbinpkgver flag needs to be added to update_chroot
and setup_board needs to be changed to pass any --getbinpkgver
flags it receives on to update_chroot.